replace cloud implementation
